safety-tool: make EXIT_AND_EMIT a field of Diagnostics
